### 2021-08-25 #3636 @csoneson
👋🏻 @mberacochea, @tkchafin, @karinlag - would (two of) you be interested in reviewing this submission for JOSS?

> Opfi: A Python package for identifying gene clusters in large genomics and metagenomics data sets
