### 2021-08-31 #3636 @csoneson
👋🏻 @afrubin - would you be interested in reviewing this submission for JOSS?

> Opfi: A Python package for identifying gene clusters in large genomics and metagenomics data sets
