### 2021-08-31 #3678 @whedon
[REVIEW]: Opfi: A Python package for identifying gene clusters in large genomics and metagenomics data sets
**Submitting author:** @alexismhill3 (<a href="http://orcid.org/0000-0002-5382-8851">Alexis Hill</a>)
**Repository:** <a href="https://github.com/wilkelab/Opfi" target ="_blank">https://github.com/wilkelab/Opfi</a>
**Version:** 0.1.0
