### 2021-08-31 #3678 @csoneson
👋🏼 @alexismhill3, @clauswilke, @Thomieh73, @afrubin - this is the review thread for the submission. All of our communications will happen here from now on.

Both reviewers have checklists at the top of this thread. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
