### 2021-09-02 #3678 @afrubin
@alexismhill3 there are several (mostly old) open issues in the GitHub repository at this time. What is the status of these issues? Some of them seem worth fixing, especially the temp file (https://github.com/wilkelab/Opfi/issues/181) and missing documentation (https://github.com/wilkelab/Opfi/issues/30) issues.

It would be helpful to know before we start adding our own comments during the review.
