### 2021-09-03 #3678 @clauswilke
@afrubin The vast majority of the old issues were stale and we have started closing them. The temp-file issue unfortunately is caused by a dependency and doesn't have an easy fix. Documentation has improved massively since spring 2020.

Any requests for improvement, new features, or bug fixes should be filed as new issues. Thanks!
