### 2021-09-07 #3678 @Thomieh73
@csoneson Hi, I am having issues with notifications from this repo. I am receiving all notifications from this entire repo. I have pushed the button to unsubscribe from this thread, but somehow I am stil receiving all notifications from this repo. Any suggestions on how to solve this.
It drives me nuts to have 100 mails in my mailbox everyday that are not relevant.
