+++ NEWS.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Opfi 0.1.1
+
+Pypi badge added to documentation.
+
 ## Opfi 0.1.0
 
 Initial release.
