### 2021-09-15 wilkelab/Opfi#198 @afrubin
PILER-CR requirement
Does the modified version of PILER-CR perform all the functions that the original version does, in addition to detecting mini CRISPR arrays or does it only detect mini CRISPR arrays? I'm assuming it's the former, but the text in the Requirements sections was ambiguous.

The current installation instructions require sudo access for PILER-CR, but looking at the Makefile it seems this is only needed to copy the executable to `/usr/local/bin/`. Users will not always have sudo access on the computer they want to run an analysis on (e.g. in an HPC environment), or they may want to avoid cluttering up their path when trying out new software. The authors should provide an alternative that doesn't require sudo access, even if that's just instructing the user to copy the executable to a directory of their choice that is in $PATH.
