### 2021-09-15 wilkelab/Opfi#199 @afrubin
Platform requirements
Based on the installation instructions, I am assuming that this program only runs on Linux or MacOS (and not Windows) but this should be clearly documented in the README.
