### 2021-09-15 wilkelab/Opfi#200 @afrubin
Installation in docs vs. README
The installation instructions and requirements in https://opfi.readthedocs.io/en/latest/installation.html and README.md do not match up. The RTD page has additional optional dependencies that aren't mentioned in the README but it does not mention the local copy of PILER-CR. The authors should make sure that there isn't conflicting installation information, either by making the text the same (and maintaining both versions) or by adding detail to the RTD page and referring to it in the README.
