### 2021-09-15 wilkelab/Opfi#201 @afrubin
Documentation for optional dependencies
The documentation at https://opfi.readthedocs.io/en/latest/installation.html#dependencies includes several optional dependencies that might be needed (or required by tests) but does not provide much extra detail.

For example, when would a user want to use GenericRepeatFinder? From what I could tell it is only mentioned on the installation page and is not cited in the paper.

Can the developers provide additional guidance in the documentation as to when someone would want to use Diamond or MMseqs2 instead of BLAST+?
