### 2021-09-15 #3678 @afrubin
I started with the installation instructions and have opened some related issues:
https://github.com/wilkelab/Opfi/issues/198
https://github.com/wilkelab/Opfi/issues/199
https://github.com/wilkelab/Opfi/issues/200
https://github.com/wilkelab/Opfi/issues/201
