### 2021-09-17 wilkelab/Opfi#201 @Thomieh73
Just a comment on GenericRepeatFinder, it has a bioconda package: https://anaconda.org/bioconda/genericrepeatfinder
you can add that to the Readthedocs installation instructions
