### 2021-09-17 wilkelab/Opfi#202 @Thomieh73
Installation using a conda environment
Hi,
I tend to install software via conda environments, to keep my system clean and that is also how I did this installation on my macbook. Maybe you can use it.

I therefore had a workaround to install opfi.

I set-up a conda environment.

```
conda create -n opfi
conda activate opfi
```

Then I install the dependencies via conda.

```
conda install -c bioconda blast=2.12.2
conda install -c bioconda diamond=2.0.11  mmseqs2=13.45111  piler-cr=1.06 genericrepeatfinder=1.0
conda install -c anaconda python
```

Then install opfi with pip3
```
pip3 install opfi
```

Then I ran the tutorial to test the installation.
