### 2021-09-20 4d7f5a5
update paths in tutorial code
