### 2021-09-20 #3678 @Thomieh73
Hi @alexismhill3,
I had created two issues when working on the installation and on your tutorial (which you already solved).
https://github.com/wilkelab/Opfi/issues/202 - which is a suggestion of how one can create a conda environment with all the dependencies.

I did not find any other problems using the tool or checking the documentation
