### 2021-09-20 #3678 @Thomieh73
@csoneson I am done with the checklist. It is an interesting tool due to the versatility of what you can research.
