### 2021-09-20 #3678 @alexismhill3
Awesome, thanks @Thomieh73!
I'm working on creating a conda package recipe for Opfi, per your suggestion, which I agree is a good idea. I think this will also resolve some other issues raised by @afrubin about the clarity of the installation process, since it will basically be condensed down to `conda install opfi`
