### 2021-09-22 #3678 @afrubin
@alexismhill3 I'm waiting to try out the new installation method, but in the meantime here are a few comments on the manuscript.

I was a little confused by the exclusive focus on metagenomics in the statement of need section. This led me to expect that the program would be analyzing some sort of shotgun metagenomic data, rather than a single genome assembly. Clearly many new genomes are arising from metagenomic studies, but this could be clearer from the start.

The implementation says that Opfi is implemented entirely in Python, but the repository does contain an optional C dependency. This should be mentioned here.

The use of capitalization in the manuscript is inconsistent (e.g. Cas1 vs. cas7, "Blastp" which is usually stylized as blastp or BLASTP). There are other formatting inconsistencies, for example Rippkea orientalis is not italicized. The authors should perform a detailed pass through the manuscript to address these kind of issues.
