### 2021-09-22 wilkelab/Opfi#205 @afrubin
"Operons"
The manuscript says that:
> the term "operon" is used interchangeably with "gene cluster" in the Opfi library

I very strongly disagree with this decision, as operon is a technical term and should be used appropriately. "gene cluster" is generic and could be used, but if a different term is needed the authors could define a new one to use in the code (a `Clump` object?).

As it stands now, I think the software is misleading and may even cause implementation issues down the line if a mode that finds true operons is added.
