### 2021-09-22 wilkelab/Opfi#206 @afrubin
Import style
As a user, I find it very unintuitive when the package I am importing doesn't have the same name as the package I installed. What was the rationale for not using `opfi` as the top level for the package? It's short, unique and seems perfectly suited.

Also, using "gene_finder" on its own could create namespace collisions - there are quite a few things with this name, and the original (to my knowledge) GeneFinder program predates Opfi by 30 years!
