### 2021-09-22 #3678 @afrubin
I have added a couple more GitHub issues arising from the manuscript:
https://github.com/wilkelab/Opfi/issues/205
https://github.com/wilkelab/Opfi/issues/206
