### 2021-09-23 7e9e9a1
make tests compatible with official piler-cr version
