### 2021-09-23 0ef0730
wrap operon diagram images to avoid overflow issues
