### 2021-09-25 ee2c180
update install instructions in README
