### 2021-09-25 ecfbbc3
update tutorial with change to recommended installation procedure
