### 2021-09-28 40b11a1
update install instructions in RTD
