+++ docs/installation.rst
@@ -6,46 +6,97 @@ Getting Started
 Installation
 ------------
 
-You can install Opfi with Pip:
+The recommended way to install Opfi is with `Bioconda <https://bioconda.github.io/>`_, which requires the `conda <https://docs.conda.io/en/latest/>`_ package manager. This will install Opfi and all of its dependencies (which you can read more about below, see :ref:`dependencies`).
+
+Currently, Bioconda supports only 64-bit Linux and Mac OS. Windows users can still install Opfi with pip (see below); however, the complete installation procedure has not been fully tested on a Windows system. 
+
+.. _install-with-conda:
+
+Install with conda (Linux and Mac OS only)
+##########################################
+
+First, set up conda and Bioconda following the `quickstart <https://bioconda.github.io/user/install.html>`_ guide. Once this is done, run:
 
 .. code-block:: bash
 
-    pip3 install opfi
+    conda install -c bioconda opfi
 
-Alternatively, you can install the latest version on Github:
+And that's it! Note that this will install Opfi in the conda environment that is currently active. To create a fresh environment with Opfi installed, do:
 
 .. code-block:: bash
 
-    git clone https://github.com/alexismhill3/Opfi.git
+    conda create --name opfi-env -c bioconda opfi
+    conda activate opfi-env
+
+.. _install-with-pip:
+
+Install with pip
+################
+
+This method does not automatically install non-Python dependencies, so they will need to be installed separately, following their individual installation instructions. A complete list of required software is provided below, see :ref:`dependencies`. Once this step is complete, install Opfi with pip by running:
+
+.. code-block:: bash
+
+    pip install opfi
+
+Install from source
+###################
+
+Finally, the latest development build may be installed directly from Github. First, non-Python :ref:`dependencies` will need to be installed in the working environment. An easy way to do this is to first install Opfi with conda using the :ref:`install-with-conda` method (we'll re-install the development version of the Opfi package in the next step). Alternatively, dependencies can be installed individually.
+
+Once dependencies have been installed in the working environment, run the following code to download and install the development build:
+
+.. code-block:: bash
+
+    git clone https://github.com/wilkelab/Opfi.git
     cd Opfi
-    pip3 install .
+    pip install . # or pip install -e . for an editable version
+    pip install -r requirements # if conda was used, this can be skipped
+
+Testing the build
+#################
+
+Regardless of installation method, users can download and run Opfi's suite of unit tests to confirm that the build is working as expected. First download the tests from Github:
+
+.. code-block:: bash
+
+    git clone https://github.com/wilkelab/Opfi
+    cd Opfi
+
+And then run the test suite using pytest:
+
+.. code-block:: bash
+
+    pytest --runslow --runmmseqs --rundiamond
+
+This may take a minute or so to complete. 
+
+.. _dependencies:    
 
 Dependencies
 ------------
 
-Opfi makes use of several third-party softwares for finding and annotating genomic features. Depending on your use case, you may not need to install all of these; however, at a minimum users should have the NCBI BLAST+ application installed in their environment. The following table provides more details about required/optional dependencies, including links to application homepages.
+Opfi uses the following bioinformatics software packages to find and annotate genomic features:
 
 .. csv-table:: 
-   :header: "Application", "Required", "Description", "Anaconda distribution"
+   :header: "Application", "Description"
 
[…]
