### 2021-09-28 073915f
finish updating install instructions and dependency guide in RTD
