### 2021-09-30 #3678 @alexismhill3
@afrubin I've created a conda recipe for Opfi and updated the documentation accordingly. These changes were merged, so you should be able to see them on the main branch now.

I had to update a few unit tests during this process. Specifically, some tests required the modified version of PILER-CR, which we have decided to deprecate the use of (see wilkelab/Opfi#207 for more details). So if you were intending to run the unit test suite, you should grab the new changes before doing so.
