+++ lib/pilercr1.06/proffrommsa.cpp
@@ -1,191 +1,0 @@
-#include "multaln.h"
-
-#define TRACE	0
-
-static void LogF(FCOUNT f)
-	{
-	if (f > -0.00001 && f < 0.00001)
-		Log("       ");
-	else
-		Log("  %5.3f", f);
-	}
-
-static const char *LocalScoreToStr(SCORE s)
-	{
-	static char str[16];
-	if (s < -1e10 || s > 1e10)
-		return "    *";
-	sprintf(str, "%5.1f", s);
-	return str;
-	}
-
-void ListProfile(const ProfPos *Prof, unsigned uLength, const MSA *ptrMSA)
-	{
-	Log("  Pos  Occ     LL     LG     GL     GG     Open  Close\n");
-	Log("  ---  ---     --     --     --     --     ----  -----\n");
-	for (unsigned n = 0; n < uLength; ++n)
-		{
-		const ProfPos &PP = Prof[n];
-		Log("%5u", n);
-		LogF(PP.m_fOcc);
-		LogF(PP.m_LL);
-		LogF(PP.m_LG);
-		LogF(PP.m_GL);
-		LogF(PP.m_GG);
-		Log("  %5.1f", -PP.m_scoreGapOpen);
-		Log("  %5.1f", -PP.m_scoreGapClose);
-		if (0 != ptrMSA)
-			{
-			const unsigned uSeqCount = ptrMSA->GetSeqCount();
-			Log("  ");
-			for (unsigned uSeqIndex = 0; uSeqIndex < uSeqCount; ++uSeqIndex)
-				Log("%c", ptrMSA->GetChar(uSeqIndex, n));
-			}
-		Log("\n");
-		}
-
-	Log("\n");
-	Log("  Pos G");
-	for (unsigned n = 0; n < g_AlphaSize; ++n)
-		Log("     %c", LetterExToChar(n));
-	Log("\n");
-	Log("  --- -");
-	for (unsigned n = 0; n < g_AlphaSize; ++n)
-		Log(" -----");
-	Log("\n");
-
-	for (unsigned n = 0; n < uLength; ++n)
-		{
-		const ProfPos &PP = Prof[n];
-		Log("%5u", n);
-		if (-1 == PP.m_uResidueGroup)
-			Log(" -", PP.m_uResidueGroup);
-		else
-			Log(" %d", PP.m_uResidueGroup);
-
-		for (unsigned uLetter = 0; uLetter < g_AlphaSize; ++uLetter)
-			{
-			FCOUNT f = PP.m_fcCounts[uLetter];
-			if (f == 0.0)
-				Log("      ");
-			else
-				Log(" %5.3f", f);
-			}
-		if (0 != ptrMSA)
-			{
-			const unsigned uSeqCount = ptrMSA->GetSeqCount();
-			Log("  ");
-			for (unsigned uSeqIndex = 0; uSeqIndex < uSeqCount; ++uSeqIndex)
-				Log("%c", ptrMSA->GetChar(uSeqIndex, n));
-			}
-		Log("\n");
-		}
-	}
-
-static void SortCounts(const FCOUNT fcCounts[], unsigned SortOrder[])
-	{
-	static unsigned InitialSortOrder[20] =
-		{
-		0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19
-		};
-	memcpy(SortOrder, InitialSortOrder, g_AlphaSize*sizeof(unsigned));
-
-	bool bAny = true;
-	while (bAny)
-		{
-		bAny = false;
[…]
