+++ lib/pilercr1.06/utils.h
@@ -1,11 +1,0 @@
-#ifndef UTILS_H
-#define UTILS_H
-
-void Quit(const char *Format, ...);
-void Warning(const char *Format, ...);
-void Log(const char *Format, ...);
-void Out(const char *Format, ...);
-char *strsave(const char *s);
-
-#endif	// UTILS_H
-

[…]
