### 2021-10-05 ac3b220
fix formatting/capitalization inconsistencies
