+++ paper/paper.md
@@ -38,11 +38,11 @@ Metagenomics is the sequencing and analysis of bacterial genomes sampled directl
 
 # Implementation
 
-Opfi is implemented entirely in Python, and can be downloaded from the Python package index and installed directly in the user’s compute environment. It consists of two major components: Gene Finder, for discovery of novel gene clusters, and Operon Analyzer, for rule-based filtering, deduplication, visualization, and re-annotation of gene clusters identified by Gene Finder. All modules generate output in a comma-separated (CSV) format that is common to the entire package.
+Opfi is implemented entirely in Python, and can be downloaded from the Python Package Index and installed directly in the user’s compute environment. It consists of two major components: Gene Finder, for discovery of novel gene clusters, and Operon Analyzer, for rule-based filtering, deduplication, visualization, and re-annotation of gene clusters identified by Gene Finder. All modules generate output in a comma-separated (CSV) format that is common to the entire package.
 
 ## Example Gene Finder usage
 
-The following example script searches for putative CRISPR-Cas loci in the Rippkaea orientalis PCC 8802 (cyanobacteria) genome. Information about the biological significance of this example, as well as data inputs and descriptions, can be found in the `tutorials` directory in the project GitHub repository. The example illustrates the use of the `Pipeline` class for setting up a gene cluster search. First, `add_seed_step` specifies a step to annotate cas1 genes, using NCBI Blastp and a database of representative Cas1 protein sequences. 10,000 bp regions directly up- and downstream of each Cas1 hit are selected for further analysis, and all other regions are discarded. Next, `add_filter_step` adds a step to annotate candidate regions for additonal cas genes. Candidates that do not have at least one additional cas gene are discarded ("filtered") from the master list of putative systems. Finally, `add_crispr_step` adds a step to search remaining candidates for CRISPR arrays, i.e. regions of alternatating ~30 bp direct repeat and variable sequences, using the PILER-CR repeat finding software [@Edgar:2007]. 
+The following example script searches for putative CRISPR-Cas loci in the *Rippkaea orientalis PCC 8802* (cyanobacteria) genome. Information about the biological significance of this example, as well as data inputs and descriptions, can be found in the `tutorials` directory in the project GitHub repository. The example illustrates the use of the `Pipeline` class for setting up a gene cluster search. First, `add_seed_step` specifies a step to annotate *cas1* genes, using protein BLAST (BLASTP) and a database of representative Cas1 protein sequences. 10,000 bp regions directly up- and downstream of each putative *cas1* gene are selected for further analysis, and all other regions are discarded. Next, `add_filter_step` adds a step to annotate candidate regions for additonal *cas* genes. Candidates that do not have at least one additional *cas* gene are discarded from the master list of putative systems. Finally, `add_crispr_step` adds a step to search remaining candidates for CRISPR arrays, i.e. regions of alternatating ~30 bp direct repeat and variable sequences, using the PILER-CR repeat finding software [@Edgar:2007]. 
 
 ```python
 from gene_finder.pipeline import Pipeline
@@ -59,13 +59,13 @@ p.add_crispr_step()
 p.run(job_id=job_id, data=genomic_data, span=10000, gzip=True)
 ```
 
-Running this code creates the CSV file "r_orientalis_results.csv", which contains information about each system identified; in this example, that is two canonical CRISPR-Cas systems, and one locus with weak homology to cas genes. Each line in the file represents a single putative feature in a candidate locus. Features from the same candidate are grouped together in the CSV. Detailed information about the output format can be found in the Opfi documentation: (<https://opfi.readthedocs.io/>).
[…]
