+++ paper/paper.bib
@@ -1,3 +1,18 @@
+@article{Fischbach:2010,
+    author = {Fischbach, Michael and Voigt, Christopher A.},
+    title = {Prokaryotic gene clusters: A rich toolbox for synthetic biology},
+    journal = {Biotechnology Journal},
+    volume = {5},
+    number = {12},
+    pages = {1277-1296},
+    keywords = {Biotechnology, Devices, Genetic parts, Refactoring, Systems biology},
+    doi = {https://doi.org/10.1002/biot.201000181},
+    url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/biot.201000181},
+    eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/biot.201000181},
+    abstract = {Abstract Bacteria construct elaborate nanostructures, obtain nutrients and energy from diverse sources, synthesize complex molecules, and implement signal processing to react to their environment. These complex phenotypes require the coordinated action of multiple genes, which are often encoded in a contiguous region of the genome, referred to as a gene cluster. Gene clusters sometimes contain all of the genes necessary and sufficient for a particular function. As an evolutionary mechanism, gene clusters facilitate the horizontal transfer of the complete function between species. Here, we review recent work on a number of clusters whose functions are relevant to biotechnology. Engineering these clusters has been hindered by their regulatory complexity, the need to balance the expression of many genes, and a lack of tools to design and manipulate DNA at this scale. Advances in synthetic biology will enable the large-scale bottom-up engineering of the clusters to optimize their functions, wake up cryptic clusters, or to transfer them between organisms. Understanding and manipulating gene clusters will move towards an era of genome engineering, where multiple functions can be “mixed-and-matched” to create a designer organism.},
+    year = {2010}
+}
+
 @article{Thomas:2012,
     author={Thomas, Torsten
     and Gilbert, Jack
