+++ paper/paper.md
@@ -30,15 +30,15 @@ bibliography: paper.bib
 
 # Summary
 
-Advances in genome sequencing technology have led to an explosion of available metagenomics data, much of which may encode novel gene systems of academic and biotechnological interest. There is a critical need for tools that can systematically and reproducibly identify gene systems in large, often low-quality metagenomics data. Therefore, we developed Opfi: a modular, rule-based pipeline for identifying functional sets of genes, such as biosynthetic gene clusters or CRISPR-Cas defense systems, in large genomics or metagenomics data sets. 
+Gene clusters perform a diverse set of functions, many of which are relevant to biotechnology, but tools for identification of gene clusters are lacking. Therefore, we developed Opfi: a modular pipeline for identification of arbitrary gene clusters in assembled genomic or metagenomic data. Opfi contains functions for de-novo annotation, de-deduplication, and visualization of putative gene clusters. It utilizes a customizable rule-based filtering approach for selection of candidate systems that adhere to user-defined criteria. Opfi is implemented in Python, and is available on the Python Package Index and the Bioconda channel.  
 
 # Statement of need
 
-Metagenomics is the sequencing and analysis of bacterial genomes sampled directly from the environment [@Thomas:2012]. This data often captures genetic diversity not present in lab-cultured organisms, and is thus a rich source of novel genetic material with potentially useful properties [@Tringe:2005]. Gene clusters, that is, sets of co-localized genes that together perform specific functions, are of particular interest to metagenomics surveyors. These systems can, for example, be leveraged to produce antimicrobial compounds or to facilitate gene editing, to name just a few applications. There are several homology search tools [@Camacho:2009; @Steinegger:2017; @Buchfink:2021] available for annotatation of singular genes (or protein domains), but these programs do not identify whole gene clusters out of the box. In many cases, researchers must combine bioinformatics tools ad-hoc, resulting in one-off pipelines that can be difficult to reproduce. Several software packages have been developed to search for specific gene clusters [@Blin:2019; @Santos-Aberturas:2019; @vanHeel:2018], but these tools may not be sufficiently flexible to identify clusters of an arbitrary genomic composition. To address these gaps, we developed a modular pipeline that integrates multiple bioinformatics tools, providing a flexible, uniform computational framework for metagenomics discovery of arbitrary gene clusters. In a recent study, we used Opfi to uncover novel CRISPR-associated transposons (CASTs) in metagenomics databases [@Rybarski:2021].
+Gene clusters are sets of co-localized, often contiguous genes that together perform specific functions. These systems have been applied successfully to meet a number of biotechnological needs, including biofuel production, organic compound synthesis, and gene editing [@Fischbach:2010].  Although there are many tools available for annotation of singular genes (or protein domains) in biological sequence data [@Camacho:2009; @Steinegger:2017; @Buchfink:2021], these programs do not identify whole gene clusters out of the box. In many cases, researchers must combine bioinformatics tools ad-hoc, resulting in one-off pipelines that can be difficult to reproduce. Several software packages have been developed to discovery of specific gene clusters [@Blin:2019; @Santos-Aberturas:2019; @vanHeel:2018], but these tools may not be sufficiently flexible to identify clusters of an arbitrary genomic composition. To address these gaps, we developed a modular pipeline that integrates multiple bioinformatics tools, providing a flexible, uniform computational framework for identification of arbitrary gene clusters. In a recent study, we used Opfi to uncover novel CRISPR-associated transposons (CASTs) in a large metagenomics dataset [@Rybarski:2021].
 
 # Implementation
 
[…]
