### 2021-10-12 369298f
Explicitly state that Opfi doesn't predict operons
