+++ paper/paper.bib
@@ -13,47 +13,6 @@
     year = {2010}
 }
 
-@article{Thomas:2012,
-    author={Thomas, Torsten
-    and Gilbert, Jack
-    and Meyer, Folker},
-    title={Metagenomics - a guide from sampling to data analysis},
-    journal={Microbial informatics and experimentation},
-    year={2012},
-    month={Feb},
-    day={09},
-    publisher={BioMed Central},
-    volume={2},
-    number={1},
-    pages={3-3},
-    abstract={Metagenomics applies a suite of genomic technologies and bioinformatics tools to directly access the genetic content of entire communities of organisms. The field of metagenomics has been responsible for substantial advances in microbial ecology, evolution, and diversity over the past 5 to 10 years, and many research laboratories are actively engaged in it now. With the growing numbers of activities also comes a plethora of methodological knowledge and expertise that should guide future developments in the field. This review summarizes the current opinions in metagenomics, and provides practical guidance and advice on sample processing, sequencing technology, assembly, binning, annotation, experimental design, statistical analysis, data storage, and data sharing. As more metagenomic datasets are generated, the availability of standardized procedures and shared data storage and analysis becomes increasingly important to ensure that output of individual projects can be assessed and compared.},
-    note={22587947[pmid]},
-    note={PMC3351745[pmcid]},
-    note={2042-5783-2-3[PII]},
-    issn={2042-5783},
-    doi={10.1186/2042-5783-2-3},
-    url={https://pubmed.ncbi.nlm.nih.gov/22587947},
-    url={https://doi.org/10.1186/2042-5783-2-3},
-    language={eng}
-}
-
-@article{Tringe:2005,
-    author={Tringe, Susannah Green
-    and Rubin, Edward M.},
-    title={Metagenomics: DNA sequencing of environmental samples},
-    journal={Nature Reviews Genetics},
-    year={2005},
-    month={Nov},
-    day={01},
-    volume={6},
-    number={11},
-    pages={805-814},
-    abstract={DNA sequencing can provide insights into organisms that are difficult to study because they are inaccessible by conventional methods such as laboratory culture.Isolation and sequencing of DNA from mixed communities of organisms (metagenomics) has revealed surprising insights into diversity and evolution.Genomic DNA sequence, and even complete genomes in some cases, has been generated from organisms that exist only in tight association with other organisms, including various obligate symbionts and pathogens, members of natural microbial consortia and an extinct cave bear.Partially assembled or unassembled genomic sequence from complex microbial communities has revealed the existence of novel and environment-specific genes.Advances in DNA sequencing technology and computational methods promise to accelerate progress in metagenomics.},
-    issn={1471-0064},
-    doi={10.1038/nrg1709},
-    url={https://doi.org/10.1038/nrg1709}
-}
-
 @article{Camacho:2009,
     author={Camacho, Christiam
     and Coulouris, George
@@ -172,6 +131,22 @@
     issn={1471-2105},
     doi={10.1186/1471-2105-8-18},
     url={https://doi.org/10.1186/1471-2105-8-18}
+}
+
+@article{Shi:2019,
+    author = {Shi, Jieming and Liang, Chun},
+    title = "{Generic Repeat Finder: A High-Sensitivity Tool for Genome-Wide De Novo Repeat Detection}",
+    journal = {Plant Physiology},
+    volume = {180},
+    number = {4},
+    pages = {1803-1815},
+    year = {2019},
+    month = {05},
[…]
