+++ paper/paper.md
@@ -38,7 +38,7 @@ Gene clusters are sets of co-localized, often contiguous genes that together per
 
 # Implementation
 
-Opfi is implemented in Python, and uses several bioinformatics tools for gene cluster feature annotation [Camacho:2009; Steinegger:2017; Buchfink:2021; Edgar:2007; Shi:2019]. Users can install Opfi and all of it's dependencies from Bioconda, using the conda package manager. Opfi consists of two major components: Gene Finder, for discovery of gene clusters, and Operon Analyzer, for rule-based filtering, deduplication, and visualization of gene clusters identified by Gene Finder. All modules generate output in a comma-separated (CSV) format that is common to the entire package.
+Opfi is implemented in Python, and uses several bioinformatics tools for gene cluster feature annotation [@Camacho:2009; @Steinegger:2017; @Buchfink:2021; @Edgar:2007; @Shi:2019]. Users can install Opfi and all of it's dependencies from Bioconda, using the conda package manager. Opfi consists of two major components: Gene Finder, for discovery of gene clusters, and Operon Analyzer, for rule-based filtering, deduplication, and visualization of gene clusters identified by Gene Finder. All modules generate output in a comma-separated (CSV) format that is common to the entire package.
 
 ## Example Gene Finder usage
 
