### 2021-10-14 #3678 @clauswilke
In brief, we have made all changes except the API changes requested in issues https://github.com/wilkelab/Opfi/issues/205 and https://github.com/wilkelab/Opfi/issues/206. We agree with these requests in principle, but they are breaking changes and that would create major problems for us in the code for a related paper (https://www.biorxiv.org/content/10.1101/2021.08.16.456562v2). This paper is about to formally appear, and we don't want to end up with a situation where this client code is immediately incompatible with the latest official release of Opfi.

We will leave these issues open and may address them at a future point.

Regarding the term "Operon" (https://github.com/wilkelab/Opfi/issues/205), we have made changes to the documentation to make it absolutely clear that we are not claiming that an `Operon` object in the library is definitely an operon in the biological meaning.
