+++ paper/paper.md
@@ -34,7 +34,7 @@ Gene clusters perform a diverse set of functions, many of which are relevant to 
 
 # Statement of need
 
-Gene clusters are sets of co-localized, often contiguous genes that together perform specific functions. These systems have been applied successfully to meet a number of biotechnological needs, including biofuel production, organic compound synthesis, and gene editing [@Fischbach:2010].  Although there are many tools available for annotation of singular genes (or protein domains) in biological sequence data [@Camacho:2009; @Steinegger:2017; @Buchfink:2021], these programs do not identify whole gene clusters out of the box. In many cases, researchers must combine bioinformatics tools ad-hoc, resulting in one-off pipelines that can be difficult to reproduce. Several software packages have been developed to discovery of specific gene clusters [@Blin:2019; @Santos-Aberturas:2019; @vanHeel:2018], but these tools may not be sufficiently flexible to identify clusters of an arbitrary genomic composition. To address these gaps, we developed a modular pipeline that integrates multiple bioinformatics tools, providing a flexible, uniform computational framework for identification of arbitrary gene clusters. In a recent study, we used Opfi to uncover novel CRISPR-associated transposons (CASTs) in a large metagenomics dataset [@Rybarski:2021].
+Gene clusters are sets of co-localized, often contiguous genes that together perform specific functions. These systems have been applied successfully to meet a number of biotechnological needs, including biofuel production, organic compound synthesis, and gene editing [@Fischbach:2010].  Although there are many tools available for annotation of singular genes (or protein domains) in biological sequence data [@Camacho:2009; @Steinegger:2017; @Buchfink:2021], these programs do not identify whole gene clusters out of the box. In many cases, researchers must combine bioinformatics tools ad-hoc, resulting in one-off pipelines that can be difficult to reproduce. Several software packages have been developed for discovery of specific gene clusters [@Blin:2019; @Santos-Aberturas:2019; @vanHeel:2018], but these tools may not be sufficiently flexible to identify clusters of an arbitrary genomic composition. To address these gaps, we developed a modular pipeline that integrates multiple bioinformatics tools, providing a flexible, uniform computational framework for identification of arbitrary gene clusters. In a recent study, we used Opfi to uncover novel CRISPR-associated transposons (CASTs) in a large metagenomics dataset [@Rybarski:2021].
 
 # Implementation
 
@@ -59,11 +59,11 @@ p.add_crispr_step()
 p.run(job_id=job_id, data=genomic_data, span=10000, gzip=True)
 ```
 
-Running this code creates the CSV file `r_orientalis_results.csv`, which contains information about each system identified; in this example, that is two canonical CRISPR-Cas systems, and one locus with weak homology to *cas* genes. Each line in the file represents a single putative feature in a candidate locus. Features from the same candidate are grouped together in the CSV. Detailed information about the output format can be found in the Opfi documentation: (<https://opfi.readthedocs.io/>).
+Running this code creates the CSV file `r_orientalis_results.csv`, which contains information about each system identified; in this example, that is two canonical CRISPR-Cas systems, and one locus with weak homology to *cas* genes. Each line in the file represents a single putative feature in a candidate locus. Features from the same candidate are grouped together in the CSV. Detailed information about the output format can be found in the Opfi [documentation](https://opfi.readthedocs.io/).
 
 ## Example Operon Analyzer usage
 
[…]
