### 2021-10-15 #3678 @afrubin
@alexismhill3 @clauswilke I understand the tension between wanting to update but avoiding breaking changes (especially with another manuscript involved), and I'm satisfied with the way you addressed the outstanding issues. I hope this software is widely-adopted and that you have the time and motivation to come back to them!

My comments on the manuscript have been addressed in full. The only outstanding issue is that you should cite Bioconda in the manuscript: https://bioconda.github.io/#citing-bioconda

FYI I was able to install Opfi and run the tests using Conda via Docker: https://docs.anaconda.com/anaconda/user-guide/tasks/docker/

I don't think you need to add this to the documentation (unless you want to support this use case) but I wanted to mention it in case it's useful for you or another user later on.
