+++ paper/paper.bib
@@ -1,3 +1,16 @@
+@article{Grüning:2018,
+    title = {{Bioconda: sustainable and comprehensive software distribution for the life sciences}},
+    author = {{Grüning}, B. and {Dale}, R. and {Sjödin}, A and {Chapman}, B. and {Rowe}, J. and {Tomkins-Tinch}, CH. and {Valieris}, R. and {Köster}, J. and {{The Bioconda Team}}},
+    year = 2018,
+    month = jul,
+    journal = {Nature Methods},
+    volume = 15,
+    number = 7,
+    pages = {475--476},
+    doi = {10.1038/s41592-018-0046-7},
+    url = {https://www.nature.com/articles/s41592-018-0046-7}
+}
+
 @article{Fischbach:2010,
     author = {Fischbach, Michael and Voigt, Christopher A.},
     title = {Prokaryotic gene clusters: A rich toolbox for synthetic biology},
