### 2021-10-15 #3678 @alexismhill3
Thanks everyone! @csoneson @afrubin I've added the Bioconda reference.
