+++ paper/paper.bib
@@ -99,7 +99,7 @@
 }
 
 @article{vanHeel:2018,
-    author = {van Heel, Auke J and de Jong, Anne and Song, Chunxu and Viel, Jakob H and Kok, Jan and Kuipers, Oscar P},
+    author = {{van Heel}, Auke J and {de Jong}, Anne and Song, Chunxu and Viel, Jakob H and Kok, Jan and Kuipers, Oscar P},
     title = "{BAGEL4: a user-friendly web server to thoroughly mine RiPPs and bacteriocins}",
     journal = {Nucleic Acids Research},
     volume = {46},
